Started work on the cards
Made a general card class to be used with the monster cards, the cards in the deck, and the special cards. Additionally, I made a class for the cards in the deck. The other two types of cards still need to be implemented, as well as reading the cards from a file.